apps/scraper-service/requirements.txt
fastapi==0.100.0
uvicorn==0.23.1
grpcio==1.57.0
grpcio-tools==1.57.0
beautifulsoup4==4.12.2
requests==2.31.0
python-multipart==0.0.6
pydantic==2.1.1
python-dotenv==1.0.0
aiohttp==3.8.5
lxml==4.9.3 